Plan 9 from Bell Labs 1993-09-04

+++ carrera/dat.h
@@ -146,7 +146,7 @@ struct KMap
 };
 
 #define	VA(k)		((k)->virt)
-#define PPN(x)		((x)>>6)
+#define PPN(x)		((ulong)(x)>>6)
 
 struct Softtlb
 {
